+++ R/AllClasses.R
@@ -14,34 +14,60 @@
 
 # Zonation curves ---------------------------------------------------------
 
-#' The Zcurves class
+#' The ZCurvesDataFrame class
 #'
-#' \code{Zcurves} class inherits class data.frame.
+#' \code{ZCurvesDataFrame} class inherits class data.frame.
 #' 
 #' Class does no implement new methods, but it is used to override some 
-#' behaviour such as plot. Usually \code{Zcurves} object belongs to a 
+#' behaviour such as plot. Usually \code{ZCurvesDataFrame} object belongs to a 
 #' \code{\link[zonator:Zresults-class]{Zresults}} object.
 #' 
-#' @note If user modifies or subsets Zcurves in a funtion it is up to the user
-#' to update the indexes in slot \code{feat.loc}
+#' @note If user modifies or subsets ZCurvesDataFrame in a funtion it is up to 
+#' the user to update the indexes in slot \code{is.feature}
 #' 
 #' @section Slots: 
 #'  \describe{
-#'    \item{\code{groups}:}{Logical indicating if the curves information is
-#'    from groups.}
 #'    \item{\code{is.feature}:}{Logical indicating whether column is actually
 #'    a feature.}
 #'  }
 #'
-#' @name Zcurves
+#' @name ZCurvesDataFrame
 #' @docType class
-#' @rdname Zcurves-class
-#' @aliases Zcurves-class
-#' @exportClass Zcurves
+#' @rdname ZCurvesDataFrame-class
+#' @aliases ZCurvesDataFrame-class
+#' @exportClass ZCurvesDataFrame
 #' @author Joona Lehtomaki <joona.lehtomaki@@gmail.com>
 #' 
 
-setClass('Zcurves', representation(groups='logical', is.feature='logical'), 
+setClass('ZCurvesDataFrame', representation(is.feature='logical'), 
+         contains = 'data.frame')
+
+#' The ZGroupCurvesDataFrame class
+#'
+#' \code{ZGroupCurvesDataFrame} class inherits class data.frame.
+#' 
+#' Class does no implement new methods, but it is used to override some 
+#' behaviour such as plot. Usually \code{ZGroupCurvesDataFrame} object belongs 
+#' to a \code{\link[zonator:Zresults-class]{Zresults}} object.
+#' 
+#' @note If user modifies or subsets ZGroupCurvesDataFrame in a funtion it is up 
+#' to the user to update the indexes in slot \code{is.group}
+#' 
+#' @section Slots: 
+#'  \describe{
+#'    \item{\code{is.group}:}{Logical indicating whether column is actually
+#'    a group.}
+#'  }
+#'
+#' @name ZGroupCurvesDataFrame
+#' @docType class
+#' @rdname ZGroupCurvesDataFrame-class
+#' @aliases ZGroupCurvesDataFrame-class
+#' @exportClass ZGroupCurvesDataFrame
+#' @author Joona Lehtomaki <joona.lehtomaki@@gmail.com>
+#' 
+
+setClass('ZGroupCurvesDataFrame', representation(is.group='logical'), 
          contains = 'data.frame')
 
 # Zonation project ------------------------------------------------------------